+++ src/features/features.service.ts
@@ -9,6 +9,7 @@ import { CreateFeatureDto } from "./dto/create-feature.dto";
 import { UpdateFeatureDto } from "./dto/update-feature.dto";
 import { UpdateFeatureOrderAndCategoryDto } from "./dto/update-feature-order-and-category.dto";
 import { GlobalService } from "../global/global.service";
+import { CustomRequest } from "../global/types/CustomRequest";
 
 @Injectable()
 export class FeaturesService {
@@ -18,7 +19,7 @@ export class FeaturesService {
     ) {}
 
     async createFeature(
-        req,
+        req: CustomRequest,
         teamId: number,
         createFeatureDto: CreateFeatureDto,
     ) {
@@ -81,6 +82,18 @@ export class FeaturesService {
 
     async findAllFeatures(teamId: number) {
         try {
+            //check for valid teamId
+            const team = await this.prisma.voyageTeam.findFirst({
+                where: {
+                    id: teamId,
+                },
+            });
+
+            if (!team) {
+                throw new NotFoundException(
+                    `TeamId (id: ${teamId}) does not exist.`,
+                );
+            }
             const allTeamFeatures = await this.prisma.projectFeature.findMany({
                 where: {
                     addedBy: {
@@ -115,12 +128,6 @@ export class FeaturesService {
                 },
                 orderBy: [{ category: { id: "asc" } }, { order: "asc" }],
             });
-
-            if (!allTeamFeatures) {
-                throw new NotFoundException(
-                    `TeamId (id: ${teamId}) does not exist.`,
-                );
-            }
 
             return allTeamFeatures;
         } catch (e) {
@@ -195,7 +202,7 @@ export class FeaturesService {
     }
 
     async updateFeatureOrderAndCategory(
-        req,
+        req: CustomRequest,
         featureId: number,
         updateOrderAndCategoryDto: UpdateFeatureOrderAndCategoryDto,
     ) {