+++ src/global/constants/formTitles.ts
@@ -2,6 +2,8 @@ export const FormTitles = {
     sprintRetroAndReview: "Retrospective & Review",
     sprintPlanning: "Sprint Planning",
     sprintCheckin: "Sprint Check-in",
+    sprintCheckinPO: "Sprint Check-in (Product Owner)",
+    sprintCheckinSM: "Sprint Check-in (Scrum Master)",
     soloProjectSubmission: "Solo Project Submission Form",
     voyageApplication: "Voyage Application Form",
     voyageProjectSubmission: "Voyage Project Submission Form",